# fix: always evaluate default values (in case it is a method)