--- a/zio-json/shared/src/test/scala/zio/json/DecoderSpec.scala
+++ b/zio-json/shared/src/test/scala/zio/json/DecoderSpec.scala
@@ -157,6 +157,33 @@ object DecoderSpec extends ZIOSpecDefault {
           assert("""{}""".fromJson[DefaultString])(isRight(equalTo(DefaultString("")))) &&
           assert("""{"s": null}""".fromJson[DefaultString])(isRight(equalTo(DefaultString(""))))
         },
+        test("dynamic default value") {
+          case class DefaultDynamic(
+            randomNumber: Double = scala.math.random(),
+            instant: java.time.Instant = java.time.Instant.now()
+          )
+
+          object DefaultDynamic {
+            implicit lazy val decoder: JsonDecoder[DefaultDynamic] = DeriveJsonDecoder.gen[DefaultDynamic]
+          }
+
+          val res1 = """{}""".stripMargin.fromJson[DefaultDynamic]
+          val res2 = """{}""".stripMargin.fromJson[DefaultDynamic]
+          val res = ZIO.fromEither {
+            for {
+              json1 <- res1
+              json2 <- res2
+            } yield (json1, json2)
+          }
+
+          for {
+            dynamics <- res
+            finalRes <-
+              assert(res1)(isRight) && assert(res2)(isRight) &&
+                assertTrue(dynamics._1.randomNumber != dynamics._2.randomNumber) &&
+                assertTrue(dynamics._1.instant != dynamics._2.instant)
+          } yield finalRes
+        },
         test("sum encoding") {
           import examplesum._
 
@@ -441,6 +468,25 @@ object DecoderSpec extends ZIOSpecDefault {
           assert(Json.Obj().as[DefaultString])(isRight(equalTo(DefaultString("")))) &&
           assert(Json.Obj("s" -> Json.Null).as[DefaultString])(isRight(equalTo(DefaultString(""))))
         },
+        test("dynamic default value") {
+          case class DefaultDynamic(
+            randomNumber: Double = scala.math.random(),
+            instant: java.time.Instant = java.time.Instant.now()
+          )
+
+          object DefaultDynamic {
+            implicit lazy val decoder: JsonDecoder[DefaultDynamic] = DeriveJsonDecoder.gen[DefaultDynamic]
+          }
+
+          for {
+            dynamics1 <- ZIO.fromEither(Json.Obj().as[DefaultDynamic])
+            _         <- ZIO.sleep(2.millis) // ensure java.time.Instant is different
+            dynamics2 <- ZIO.fromEither(Json.Obj().as[DefaultDynamic])
+          } yield assertTrue(
+            dynamics1.randomNumber != dynamics2.randomNumber,
+            dynamics1.instant != dynamics2.instant
+          )
+        } @@ TestAspect.withLiveClock,
         test("aliases") {
           import exampleproducts._