This PR attempts to fix the same as #1119 but as simple as possible without annotations. Credits to @Andrapyre, author of #1119, who made some magnolia changes in May 2024 to add `evaluateDefault`. I guess in preparation for usage, among others, in zio-json.
I do not know why his PR in zio-json was not picked up. It went quite stale. Hopefully this PR or the old PR can be picked up because it seems like an important fix.

Closes #1055

Note: I do not claim the bounty as prework was already done but it is unclear if the work was rejected, failed or just never looked at.