@ThijsBroersen 2025-01-22
Now let's hope nobody is depending on memoized default method values as a feature 🙈