@Andrapyre 2025-01-22
@ThijsBroersen , I was disappointed to see that https://github.com/zio/zio-json/pull/1119 never received a review. I don't know why - I asked multiple people multiple times and then left it as is. Now it looks like it would need some rebasing to get it ready for a merge. This PR has the advantage of being less complex, whereas https://github.com/zio/zio-json/pull/1119 should be more performant in general, since all non-dynamic defaults will be evaluated at compile time, instead of at runtime. I'm happy to rebase https://github.com/zio/zio-json/pull/1119 if a member of the zio-json team confirms that the solution is a go-ahead. Otherwise, best of luck in getting this across the finish line :)