@ThijsBroersen 2025-01-22
@Andrapyre perhaps it would be interesting to see if there is much performance difference. Would evaluating a constant really be a big penalty?
I think the ideal solution would be if magnolia would indicate if the value is a constant or not.