@plokhotnyuk 2025-01-23
I will remeasure impact of dynamic evaluation of default with benchmarks and will share report here.