@plokhotnyuk 2025-01-23
Below are results for real-world message samples using JDK-21 on  i9-13900K CPU.

### Scala 3

#### Before
```scala
Benchmark                             (size)   Mode  Cnt        Score        Error  Units
GeoJSONReading.zioJson                   N/A  thrpt   15     9465.550 ±    165.541  ops/s
GitHubActionsAPIReading.zioJson          N/A  thrpt   15   310255.930 ±   4598.464  ops/s
GoogleMapsAPIReading.zioJson             N/A  thrpt   15    19164.603 ±    172.197  ops/s
OpenRTBReading.zioJson                   N/A  thrpt   15   164941.515 ±   1186.926  ops/s
TwitterAPIReading.zioJson                N/A  thrpt   15    17456.399 ±    721.673  ops/s
```

#### After
```scala
Benchmark                             (size)   Mode  Cnt        Score        Error  Units
GeoJSONReading.zioJson                   N/A  thrpt   15     9545.312 ±    192.893  ops/s
GitHubActionsAPIReading.zioJson          N/A  thrpt   15   320703.547 ±   3579.725  ops/s
GoogleMapsAPIReading.zioJson             N/A  thrpt   15    19379.795 ±    182.351  ops/s
OpenRTBReading.zioJson                   N/A  thrpt   15   131212.485 ±    816.589  ops/s
TwitterAPIReading.zioJson                N/A  thrpt   15    16277.179 ±    472.626  ops/s
``` 

### Scala 2

TBD