@plokhotnyuk 2025-01-23
Below are results for real-world message samples using JDK-21 on  i9-13900K CPU.

### Scala 3

#### Before
```scala
Benchmark                             (size)   Mode  Cnt        Score        Error  Units
GeoJSONReading.zioJson                   N/A  thrpt   15     9465.550 ±    165.541  ops/s
GitHubActionsAPIReading.zioJson          N/A  thrpt   15   310255.930 ±   4598.464  ops/s
GoogleMapsAPIReading.zioJson             N/A  thrpt   15    19164.603 ±    172.197  ops/s
OpenRTBReading.zioJson                   N/A  thrpt   15   164941.515 ±   1186.926  ops/s
TwitterAPIReading.zioJson                N/A  thrpt   15    17456.399 ±    721.673  ops/s
```

#### After
```scala
Benchmark                             (size)   Mode  Cnt        Score        Error  Units
GeoJSONReading.zioJson                   N/A  thrpt   15     9545.312 ±    192.893  ops/s
GitHubActionsAPIReading.zioJson          N/A  thrpt   15   320703.547 ±   3579.725  ops/s
GoogleMapsAPIReading.zioJson             N/A  thrpt   15    19379.795 ±    182.351  ops/s
OpenRTBReading.zioJson                   N/A  thrpt   15   131212.485 ±    816.589  ops/s
TwitterAPIReading.zioJson                N/A  thrpt   15    16277.179 ±    472.626  ops/s
``` 

### Scala 2

#### Before
```scala
Benchmark                         Mode  Cnt       Score      Error  Units
GeoJSONReading.zioJson           thrpt   15   10713.393 ±  179.807  ops/s
GitHubActionsAPIReading.zioJson  thrpt   15  263676.440 ± 8420.895  ops/s
GoogleMapsAPIReading.zioJson     thrpt   15   15986.893 ±  245.732  ops/s
OpenRTBReading.zioJson           thrpt   15  141819.438 ±  402.495  ops/s
TwitterAPIReading.zioJson        thrpt   15   10187.695 ±  742.839  ops/s
```

#### After
```scala
Benchmark                         Mode  Cnt       Score      Error  Units
GeoJSONReading.zioJson           thrpt   15   10698.177 ±  123.256  ops/s
GitHubActionsAPIReading.zioJson  thrpt   15  255979.354 ± 8965.585  ops/s
GoogleMapsAPIReading.zioJson     thrpt   15   15300.128 ±  350.823  ops/s
OpenRTBReading.zioJson           thrpt   15  134179.618 ±  700.721  ops/s
TwitterAPIReading.zioJson        thrpt   15   10204.254 ±  870.285  ops/s
``` 

The most affected is `OpenRTBReading` which is based on nested product types with _all_ fields have own default values defined.