@ThijsBroersen 2025-01-23
@plokhotnyuk could you also show the flame graph for the Before results?