@Andrapyre 2025-01-23
@plokhotnyuk , if it's not too much work, would be interesting to see benchmarks for https://github.com/zio/zio-json/pull/1119 and whether it solves this regression.