@ThijsBroersen 2025-01-23
@plokhotnyuk thanks for running the benchmarks